+++ src/main/java/ar/edu/unq/desapp/grupoa/model/Menu.java
@@ -4,13 +4,13 @@ import java.time.LocalTime;
 import java.util.GregorianCalendar;
 import java.util.List;
 
-import ar.edu.unq.desapp.grupoa.model.Exception.DataIncompleteException;
-import ar.edu.unq.desapp.grupoa.model.Exception.DescriptionLengthException;
-import ar.edu.unq.desapp.grupoa.model.Exception.EmptyListException;
-import ar.edu.unq.desapp.grupoa.model.Exception.EmptyStringException;
-import ar.edu.unq.desapp.grupoa.model.Exception.IrrationalAmountException;
-import ar.edu.unq.desapp.grupoa.model.Exception.IrrationalPriceException;
-import ar.edu.unq.desapp.grupoa.model.Exception.NameLengthException;
+import ar.edu.unq.desapp.grupoa.model.exceptions.DataIncompleteException;
+import ar.edu.unq.desapp.grupoa.model.exceptions.DescriptionLengthException;
+import ar.edu.unq.desapp.grupoa.model.exceptions.EmptyListException;
+import ar.edu.unq.desapp.grupoa.model.exceptions.EmptyStringException;
+import ar.edu.unq.desapp.grupoa.model.exceptions.IrrationalAmountException;
+import ar.edu.unq.desapp.grupoa.model.exceptions.IrrationalPriceException;
+import ar.edu.unq.desapp.grupoa.model.exceptions.NameLengthException;
 import lombok.Getter;
 import lombok.NonNull;
