+++ src/main/java/ar/edu/unq/desapp/grupoa/model/exceptions/DescriptionLengthException.java
@@ -1,4 +1,4 @@
-package ar.edu.unq.desapp.grupoa.model.Exception;
+package ar.edu.unq.desapp.grupoa.model.exceptions;
 
 public class DescriptionLengthException extends RuntimeException {
     public DescriptionLengthException(String message) {
