+++ src/main/java/ar/edu/unq/desapp/grupoa/model/exceptions/IrrationalAmountException.java
@@ -1,4 +1,4 @@
-package ar.edu.unq.desapp.grupoa.model.Exception;
+package ar.edu.unq.desapp.grupoa.model.exceptions;
 
 public class IrrationalAmountException extends RuntimeException {
 	public IrrationalAmountException (String message) {
