+++ src/main/java/ar/edu/unq/desapp/grupoa/model/exceptions/IrrationalPriceException.java
@@ -1,4 +1,4 @@
-package ar.edu.unq.desapp.grupoa.model.Exception;
+package ar.edu.unq.desapp.grupoa.model.exceptions;
 
 public class IrrationalPriceException extends RuntimeException {
 	public IrrationalPriceException(String message) {
