+++ src/main/java/ar/edu/unq/desapp/grupoa/model/exceptions/NameLengthException.java
@@ -1,4 +1,4 @@
-package ar.edu.unq.desapp.grupoa.model.Exception;
+package ar.edu.unq.desapp.grupoa.model.exceptions;
 
 public class NameLengthException extends RuntimeException {
     public NameLengthException(String message) {
